change function for generate slug; repository

+++ source/app/routes/api/deletePostWithoutRedirect.ts
@@ -1,8 +1,10 @@
 import { getSession } from "~/shared/.server/services/session";
 
 import type { Route } from "../../+types/root";
-import { SESSION_USER_KEY } from "~/shared/constants/common";
+import { ROLE_ADMIN, SESSION_USER_KEY } from "~/shared/constants/common";
 import { deletePostById } from "~/shared/.server/repository/posts";
+import { getUserById } from "~/shared/.server/repository/users";
+import { errorHandler } from "~/shared/.server/utils/errorHandler";
 
 export async function action({ request }: Route.ActionArgs) {
   const formData = await request.formData();
@@ -16,6 +18,24 @@ export async function action({ request }: Route.ActionArgs) {
 
   const session = await getSession(request.headers.get("cookie"));
   const sessionUser = session.get(SESSION_USER_KEY);
+  // if(!sessionUser) //TODO
 
-  await deletePostById(Number(postId), sessionUser.id);
+  try {
+    const existingUser = await getUserById(sessionUser.id);
+
+    if (!existingUser) {
+      throw new Error("User with such id does not exist");
+    }
+
+    if (
+      existingUser.id !== sessionUser.id &&
+      existingUser.role !== ROLE_ADMIN
+    ) {
+      throw new Error("Access denied"); //TODO
+    }
+
+    await deletePostById(Number(postId), sessionUser.id);
+  } catch (error) {
+    errorHandler(error);
+  }
 }
